Move Warning Dialogs out of Catalog, and into their own class.
